first implementation of full design pattern with mockup test MineField spin-up
Working on a derived MPC/MVI pattern where the controller is the Game itself (spawned when starting a new game), is the sole owner of View and Mode (which I call State) and is the sole point of contact to access the game's functionality.

The view is simplified into just a SWidget, decoupled from the settings panel, it will most likely just get the state and rebuild itself at every step

A lot of things to clean but the first tests are promising.